#72 feat: 홈 todayStudyTime Long -> Int

+++ core/data/src/main/java/com/ds/studify/core/data/dto/response/ResponseHomeDto.kt
@@ -9,7 +9,7 @@ data class ResponseHomeDto(
     @SerialName("nickName")
     val nickName: String,
     @SerialName("todayStudyTime")
-    val todayStudyTime: Long
+    val todayStudyTime: Int
 ) {
     fun toEntity() = HomeEntity(
         nickName = this.nickName,
