+++ core/domain/src/main/java/com/ds/studify/core/domain/entity/HomeEntity.kt
@@ -2,5 +2,5 @@ package com.ds.studify.core.domain.entity
 
 data class HomeEntity(
     val nickName: String,
-    val todayStudyTime: Long
+    val todayStudyTime: Int
 )
